Move `AsUnderscore` into `Casts` lint pass
